+++ integrations/github/main.py
@@ -308,6 +308,11 @@ async def resync_teams(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
                         ),
                     )
 
+                if selector.include_external_group:
+                    teams = await RestTeamExporter(
+                        rest_client
+                    ).enrich_teams_with_external_group(teams, org_name)
+
                 yield teams