+++ integrations/github/CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
-- Added a `sinceDate` pull request selector to ingest all closed PRs since an absolute date.
+- Added a `closedSinceDate` pull request selector to ingest all closed PRs since an absolute date.
 
 
 ## 6.0.3 (2026-06-15)